+++ modred/tests/testbpod.py
@@ -214,7 +214,8 @@ class TestBPODHandles(unittest.TestCase):
             self.assertEqual(v, data_members_modified[k])
         
         max_vecs_per_node = 500
-        my_BPOD = BPODHandles(my_IP, max_vecs_per_node=max_vecs_per_node, verbosity=0)
+        my_BPOD = BPODHandles(
+            my_IP, max_vecs_per_node=max_vecs_per_node, verbosity=0)
         data_members_modified = copy.deepcopy(data_members_default)
         data_members_modified['vec_space'].max_vecs_per_node = \
             max_vecs_per_node
@@ -237,25 +238,27 @@ class TestBPODHandles(unittest.TestCase):
             np.random.random, ((num_vecs, num_vecs)))
         L_sing_vecs_true, sing_vals_true, R_sing_vecs_true = \
             _parallel.call_and_bcast(util.svd, Hankel_mat_true)
+
         my_BPOD = BPODHandles(None, verbosity=0)
-        _parallel.barrier()
+        my_BPOD.Hankel_mat = Hankel_mat_true
+        my_BPOD.sing_vals = sing_vals_true
+        my_BPOD.L_sing_vecs = L_sing_vecs_true
+        my_BPOD.R_sing_vecs = R_sing_vecs_true
 
         L_sing_vecs_path = join(test_dir, 'L_sing_vecs.txt')
         R_sing_vecs_path = join(test_dir, 'R_sing_vecs.txt')
         sing_vals_path = join(test_dir, 'sing_vals.txt')
         Hankel_mat_path = join(test_dir, 'Hankel_mat.txt')
-        my_BPOD.Hankel_mat = Hankel_mat_true
-        my_BPOD.sing_vals = sing_vals_true
-        my_BPOD.L_sing_vecs = L_sing_vecs_true
-        my_BPOD.R_sing_vecs = R_sing_vecs_true
-        
         my_BPOD.put_decomp(L_sing_vecs_path, sing_vals_path, R_sing_vecs_path)
         my_BPOD.put_Hankel_mat(Hankel_mat_path)
+        _parallel.barrier()
+
         BPOD_load = BPODHandles(None, verbosity=0)
         
         BPOD_load.get_decomp(
             L_sing_vecs_path, sing_vals_path, R_sing_vecs_path)
-        Hankel_mat_loaded = _parallel.call_and_bcast(util.load_array_text, Hankel_mat_path)
+        Hankel_mat_loaded = _parallel.call_and_bcast(
+            util.load_array_text, Hankel_mat_path)
 
         np.testing.assert_allclose(Hankel_mat_loaded, Hankel_mat_true)
         np.testing.assert_allclose(BPOD_load.L_sing_vecs, L_sing_vecs_true)